+++ CHANGES.md
@@ -1,4 +1,4 @@
-### 2.2.2.dev0 - In development
+### 2.2.2 - 2018-08-30 - Bug fix release
 
 - Add a `set_start_method` function in `loky.backend.context`. Note
   that now, `loky` does not respect the start method set using